Add utilities router
