+++ backend/routes/api/apiPartidos.js
@@ -2,29 +2,14 @@
 const express = require('express');
 const router = express.Router();
 const _ = require('lodash');
+const util = require('../utilities');
+const queryPage = util.queryPage;
+const sendRes = util.sendRes;
+
 const Partido = require('../../models/partido');
 const Marcador = require('../../models/marcador');
 const Equipo = require('../../models/equipo');
 const TipoEvento = require('../../models/tipoEvento.js');
-
-// Solo con el objetivo de enviar siempre una misma respuesta
-function sendRes(res, cod, data, message, error) {
-  res.status(cod);
-  return res.json({ data, message, error });
-}
-
-// Normalizar parametros para el paginado
-function queryPage(req, res, next) {
-  // en caso de no estar definido se fuersa a 0
-  const skip = _.get(req, 'query.skip', 0) || 0;
-  // en caso de no estar definido se fuersa a 15
-  let limit = _.get(req, 'query.limit', 15) || 15;
-  limit = parseInt(limit, 10);
-  req.query.skip = parseInt(skip, 10);
-  req.query.limit = limit > 0 ? limit : 15;
-  // Continuar con la consulta ala API
-  next();
-}
 
 // Recupera todos los partidos
 // http://localhost:3000/api/partido/?skip=1&limit=1
